configparser is your friend.
I have another suggestion, but I have no idea if you like it, because it changes the config.txt format a bit.

The Python standard library offers a module named configparser, which can read and write "win.ini" style files. The keys in this file, as well as the values, can contain spaces, but unlike your config, a "=" instead of ":" is used to separate keys and values.

Like in your config, sections can be set up in the file, but the syntax again is slightly different: section names need to be written in [these brackets].

Another change I made was to .lower() the language name in the beginnning after argparse, so all the calls to .lower() in the later code are no longer required.